fix: resolve all clippy warnings for CI
- Add #[allow(dead_code)] for unused but reserved functions/fields
- Replace manual prefix stripping with .strip_prefix()
- Add #[allow(clippy::too_many_arguments)] where needed
- Add #[allow(clippy::should_implement_trait)] for from_str/default methods
- Add #[allow(clippy::type_complexity)] for complex SQL aggregate types
- Fix &PathBuf → &Path in build/clone/compose functions
- Remove redundant .map(|x| x) identity transforms
- Fix identical if/else branches in teams/mod.rs
- cargo fmt applied to all files

cargo clippy -- -D warnings: 0 errors
cargo fmt --check: clean

+++ src/engine/disk_monitor.rs
@@ -48,7 +48,7 @@ impl DiskStats {
             let stat = unsafe { stat.assume_init() };
 
             // Calculate sizes
-            let block_size = stat.f_frsize as u64;
+            let block_size = stat.f_frsize;
             let total_bytes = stat.f_blocks as u64 * block_size;
             let free_bytes = stat.f_bfree as u64 * block_size;
             let available_bytes = stat.f_bavail as u64 * block_size;
